refactor: extract PanelCoordinator from MainWindowView.swift
Extract panel routing, slot rendering, and workspace views from the
1,928-line MainWindowView.swift into PanelCoordinator.swift. No logic
changes — only code movement and access modifier adjustments.
